requirements.txt
aiogram==3.11.0
Jinja2==3.1.3
python-dotenv==1.0.1
weasyprint==52.5
yookassa==2.4.0
apscheduler==3.10.4
python-docx==1.1.0